fix: one renderer for the dev-mode command, and quote the sync ids too
Two review findings, both extending the last commit to where it stopped short.

`create-branch --dev-mode` is a THIRD site handing over `cube data-model dev-mode` — and
the one where the branch comes from the caller's own argv rather than a server payload, so
`create-branch 42 'feat#1234' --dev-mode` printed a command that pastes as `dev-mode 42
feat` and a comment. It now shares the renderer, which moved from `deployments.rs` to
`data_model.rs`: the module that owns the subcommand should own how it is spelled, and
with all three going through it the quoting test speaks for all three rather than two.

The other finding asked for a decision rather than a patch: `sync_job_id` is user-supplied
at `status`/`result`/`cancel`, with no `value_parser` constraining it, so `cube dbt result
42 'j;whoami'` handed back a command that pastes as two. The convention is now every
user-supplied value inside a suggested command, and `shell_quote`'s doc says so. Sync ids
are opaque hex in practice and a wrong one 404s loudly — but branch names were "usually
fine" too until `#1234-fix` turned up, and a rule with an exemption list is one somebody
has to re-derive at the next call site.

Two pre-existing sites outside this branch — `cube login --context {name}` in `main.rs`
and `context.rs` — still interpolate unquoted. Left alone deliberately: unrelated code,
and widening the diff to reach them would bury what this PR is for.

52 tests, clippy and fmt clean.
